pyproject.toml
[project]
name = "dagster_databricks_components_demo"
requires-python = ">=3.9,<=3.13.3"
version = "0.1.0"
dependencies = [
    "dagster==1.11.2",
    "databricks-sdk>=0.58.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
    "ruff>=0.12.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "dagster_databricks_components_demo"
registry_modules = [
    "dagster_databricks_components_demo.components.*",
]


